requirements.txt
bs4==0.0.2
py-cord
pydotenv==0.0.7
requests==2.32.3
uptime_kuma_api==1.2.1
python-dotenv==1.0.1
lxml==5.3.0